requirements.txt
# A2A Agent Service - 核心依赖文件
# 支持 SQLite 和 PostgreSQL 两种数据库

# 核心 Web 框架
fastapi>=0.115.2
uvicorn[standard]==0.24.0
websockets==12.0

# 数据库支持
sqlalchemy==2.0.23
aiosqlite==0.19.0  # SQLite 异步支持
psycopg2-binary==2.9.9; platform_system != "Windows"  # PostgreSQL (可选)
psycopg2==2.9.9; platform_system == "Windows"  # PostgreSQL Windows (可选)
alembic==1.12.1

# 消息队列
celery==5.3.4
redis==5.0.1

# HTTP 客户端
httpx>=0.28.1
aiohttp==3.9.1
requests>=2.31.0  # 兼容性支持

# MCP 协议支持
mcp>=0.1.0  # Model Context Protocol (如果有官方包)
jsonrpc-base>=2.2.0  # JSON-RPC协议支持
jsonrpc-async>=2.1.1  # 异步JSON-RPC客户端

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据模型与验证
pydantic
pydantic-settings

# LLM 集成
openai
# 替换anthropic为智谱AI官方SDK
zai-sdk
# anthropic  # 已替换为zai-sdk

# 开发和调试工具
python-dotenv==1.0.0
loguru==0.7.2
pydantic-core
pytest
pytest-asyncio
black
flake8

# A2A Protocol (可选 - 如果需要完整SDK支持)
a2a-sdk
